[Refactor] #26 - UI 공통 객체(TableViewCell) 리팩토링
- SearchLogCustomView를 삭제하였습니다.
- Cell 내 객체로 리팩토링하였습니다.
