+++ WeatherWhat/WeatherWhat/View/LocationHistoryCell.swift
@@ -9,7 +9,8 @@ import UIKit
 
 class LocationHistoryCell: UITableViewCell {
 
-    let searchLogCustomView = SearchLogCustomView()
+    let label = UILabel()
+    let cancelButton = UIButton()
 
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
@@ -22,17 +23,32 @@ class LocationHistoryCell: UITableViewCell {
     }
 
     func configureUI() {
-        self.contentView.addSubview(searchLogCustomView)
         self.backgroundColor = .clear
+
+        [label, cancelButton].forEach {
+            self.addSubview($0)
+        }
+
+        label.font = .suit(.regular, size: 13)
+        label.textColor = .pureBlack
+
+        cancelButton.setImage(UIImage(named: "deleteIcon"), for: .normal)
     }
 
-    func setConstraints() {
-        searchLogCustomView.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
+    private func setConstraints() {
+        label.snp.makeConstraints { make in
+            make.leading.equalToSuperview().offset(14)
+            make.centerY.equalToSuperview()
+        }
+
+        cancelButton.snp.makeConstraints { make in
+            make.trailing.equalToSuperview().offset(-14)
+            make.centerY.equalToSuperview()
+            make.size.equalTo(10)
         }
     }
 
     func configureCell(data: LocationHistory, indexPath: IndexPath) {
-        searchLogCustomView.label.text = data.history[indexPath.row].address
+        label.text = data.history[indexPath.row].address
     }
 }
